backend/requirements.txt
Flask
flask-mysql
mysql-connector-python
flask-cors